# packages/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FasterClaw API",
    "version": "1.0.0",
    "description": "API for managing OpenClaw instances and subscriptions"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.fasterclaw.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CheckoutResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "url"
        ]
      },
      "CreateCheckoutRequest": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "starter",
              "pro",
              "enterprise"
            ]
          }
        },
        "required": [
          "plan"
        ]
      },
      "PortalResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "url"
        ]
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "paidAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "invoiceUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "invoicePdf": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          }
        },
        "required": [
          "id",
          "amount",
          "status",
          "createdAt",
          "paidAt",
          "invoiceUrl",
          "invoicePdf"
        ]
      },
      "InvoiceList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Invoice"
        }
      },
      "Subscription": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid"
          },
          "userId": {
            "type": "string",
            "format": "cuid"
          },
          "stripeCustomerId": {
            "type": "string"
          },
          "stripeSubscriptionId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "plan": {
            "type": "string",
            "nullable": true,
            "enum": [
              "starter",
              "pro",
              "enterprise"
            ]
          },
          "instanceLimit": {
            "type": "number"
          },
          "currentPeriodStart": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "currentPeriodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "cancelAtPeriodEnd": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "stripeCustomerId",
          "stripeSubscriptionId",
          "status",
          "plan",
          "instanceLimit",
          "currentPeriodStart",
          "currentPeriodEnd",
          "cancelAtPeriodEnd",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlanConfig": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "priceId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "instanceLimit": {
            "type": "number"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "priceId",
          "price",
          "instanceLimit",
          "features"
        ]
      },
      "Plans": {
        "type": "object",
        "properties": {
          "starter": {
            "$ref": "#/components/schemas/PlanConfig"
          },
          "pro": {
            "$ref": "#/components/schemas/PlanConfig"
          },
          "enterprise": {
            "$ref": "#/components/schemas/PlanConfig"
          }
        }
      },
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/Subscription"
          },
          "plans": {
            "$ref": "#/components/schemas/Plans"
          }
        },
        "required": [
          "subscription",
          "plans"
        ]
      },
      "WebhookResponse": {
        "type": "object",
        "properties": {
          "received": {
            "type": "boolean"
          }
        },
        "required": [
          "received"
        ]
      }
    },
...
  },
  "paths": {
...
    "/billing/checkout": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create Stripe Checkout session for subscription",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/portal": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Create Stripe Customer Portal session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Portal session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/invoices": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Get invoice history from Stripe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of invoices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/subscription": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Get current subscription status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/webhook": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Handle Stripe webhook events",
        "responses": {
          "200": {
            "description": "Webhook received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}